+++ libsel4/sel4_plat_include/stm32mp2/sel4/plat/api/constants.h
@@ -1,0 +1,14 @@
+/*
+ * Copyright 2026, STMicroelectronics
+ *
+ * SPDX-License-Identifier: BSD-2-Clause
+ */
+
+#pragma once
+
+#include <sel4/config.h>
+#include <sel4/arch/constants_cortex_a35.h>
+
+#ifdef CONFIG_ARCH_AARCH32
+#error "AARCH32 is unsupported"
+#endif